% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adddata.R
\name{adddata}
\alias{adddata}
\title{Data frame
Description of the spruce data frame.}
\usage{
adddata(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
The head of the data frame.
}
\description{
Data frame
Description of the spruce data frame.
}
