% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBinFunc.R
\name{mybin}
\alias{mybin}
\title{MyBinFunc}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{how many iterations}

\item{n}{test}

\item{p}{probability}
}
\value{
vector containing probability estimates of 0 to n successes
}
\description{
MyBinFunc
}
\examples{
\dontrun{bin_100 <- mybin(100, 10, 0.7)}

}
