% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R, R/rpackage.R
\name{myci}
\alias{myci}
\title{Title calculate 95 confidence interval for mean}
\usage{
myci(x)

myci(x)
}
\arguments{
\item{x}{a sample}
}
\value{
confidence interval for mean

confidence interval for mean
}
\description{
Title calculate 95 confidence interval for mean

Title calculate 95 confidence interval for mean
}
