% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myclt.R
\name{myclt}
\alias{myclt}
\title{mycltfunc}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{Sample size}

\item{iter}{iterations}
}
\value{
- histogram
}
\description{
mycltfunc
}
\examples{
\dontrun{myclt(n = 30, iter = 1000)}
}
