% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R, R/rpackage.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{to find the Max Likelihood}
\usage{
mymaxlikg(lfun = "logbin2", theta)

mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{logbin2 function}

\item{theta}{A vector of parameter value}
}
\value{
The parameter coresponding to the maximum likelihood

The parameter coresponding to the maximum likelihood
}
\description{
to find the Max Likelihood

to find the Max Likelihood
}
