% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MynCurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper bound shading}
}
\value{
list
}
\description{
Title
}
