% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of available seats}

\item{gamma}{The acceptable probability of exceeding the number of available seats}

\item{p}{Probability of a ticket holder showing up}
}
\value{
named list containing nd, nc, N, p and gamma
}
\description{
Title
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
