% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{scatter histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{x variable as a vector}

\item{y}{y variable as a vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
a plot
}
\description{
scatter histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
