% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{Spruce Data}
\format{
A data frame with [number of rows] rows and [number of columns] variables.
}
\source{
Specify the source of the data, if applicable.
}
\usage{
spruce
}
\description{
Description of the spruce data frame.
}
\examples{
# Load the spruce data
data(spruce)

# Display the head of the data
head(spruce)

}
\seealso{
\code{\link{head}}
}
\keyword{datasets}
