% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R, R/rpackage.R
\name{myncurve}
\alias{myncurve}
\title{This function plots the normal density curve and calculates the probability of X <= a.}
\usage{
myncurve(mu, sigma, a, x)

myncurve(mu, sigma, a, x)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{Alpha value}

\item{x}{Additional x-value}
}
\value{
A list containing the plot and the probability P(X <= a).

A list containing the plot and the probability P(X <= a).
}
\description{
This function plots the normal density curve and calculates the probability of X <= a.

This function plots the normal density curve and calculates the probability of X <= a.
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
myncurve(mu = 10, sigma = 5, a = 6)
}
